Refactor python

+++ PerudoBot.Discord/views/game_setup.py
@@ -1,15 +1,14 @@
 import discord
-from cogs.models.round import Round
-from cogs.models.setup import GameSetup
-from cogs.utils.client import GameClient
-
+from models import Round, GameSetup
+from utils import GameClient
 
 class GameSetupView(discord.ui.View):
     def __init__(self, game_client):
-        super().__init__()
+        super().__init__(timeout=600)
         self.game_client : GameClient = game_client
         self.player_count = 0
         self.first_round : Round = None
+        self.timed_out = True
 
     @discord.ui.button(label='Join', style=discord.ButtonStyle.gray)
     async def add(self, interaction: discord.Interaction, button: discord.ui.Button):
@@ -54,6 +53,7 @@ class GameSetupView(discord.ui.View):
         if self.player_count < 2:
             await interaction.response.send_message("Need at least two players to start", ephemeral=True)
         else:
+            self.timed_out = False
             self.clear_items()
             await interaction.response.edit_message(view=self)
             self.stop()
@@ -63,11 +63,16 @@ class GameSetupView(discord.ui.View):
         bot_user = await interaction.guild.fetch_member('743151009689501818')
         response = self.game_client.add_player(bot_user.id, bot_user.display_name, is_bot=bot_user.bot)
         
+        if not response.is_success:
+            await interaction.response.send_message(response.error_message, ephemeral=True)
+            return
+        
         game_setup = GameSetup(response.data)
         self.player_count += 1
+        button.disabled = True
 
         if interaction.channel.last_message_id == interaction.message.id:
-            await interaction.response.edit_message(embed=GameSetupEmbed(game_setup))
+            await interaction.response.edit_message(view=self, embed=GameSetupEmbed(game_setup))
         else:
             await interaction.message.delete()
             await interaction.channel.send(view=self, embed=GameSetupEmbed(game_setup))
